+++ src/app/BasePage.tsx
@@ -2,7 +2,7 @@ import React, {Suspense} from 'react'
 import {Route, Redirect, Switch} from 'react-router-dom'
 import {LayoutSplashScreen} from '../common/layout'
 import DashboardPage from './pages/Dashboard'
-import SettingPage from './pages/Contact/ContactSettings'
+import SettingPage from './pages/Contact/contactSettings'
 import {Contacts} from './pages/Contact'
 import {UserAccountsPage} from './pages/UserAccount'
 import {LendingPage} from './pages/LendingPage'
