+++ src/app/pages/Contact/contactAdd.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import CompanyForm from './CompanyForm'
+import CompanyForm from './companyForm'
 import IndividualForm from './IndividualForm'
 
 // add country later when the module is implemented
