+++ src/redux/rootReducer.ts
@@ -3,12 +3,12 @@ import { combineReducers } from 'redux'
 
 import * as auth from '../app/modules/Auth/_redux/authRedux'
 import { userAccountsSlice } from './userAccounts/userAccountsSlice'
-import { contactsSlice } from '../app/pages/Contact/state/ContactsSlice'
+import { contactsSlice } from '../app/pages/Contact/state/contactsSlice'
 import { accountsSlice } from './accounts/accountsSlice'
 import { countriesSlice } from './countries/countriesSlice'
 import { currenciesSlice } from './currencies/currenciesSlice'
 import { iconsSlice } from './icons/iconsSlice'
-import { contactMethodsSlice } from '../app/pages/Contact/ContactMethods/state/ContactMethodsSlice'
+import { contactMethodsSlice } from '../app/pages/Contact/ContactMethods/state/contactMethodsSlice'
 import { platformsSlice } from '../app/pages/Platform/state/platformsSlice'
 import { originatorsSlice } from '../app/pages/Originator/state/originatorsSlice'
